First working and interactive version of collision app
In this commit, the first interactive version of the collision app is
submitted. In this version, the user can use the interactivity of the
app to play with the colliding balls. In addition, the balls are
colliding according to the laws of physics that govern both the elastic
and inelastic collision